fast shift subtract stack

+++ lib/sgbm_helper.hpp
@@ -43,8 +43,8 @@ inline float hamming_dist(unsigned long long a,
 unsigned long long b);
 
 
-extern "C" void shift_subtract_stack(unsigned long long int * L,
-unsigned long long int * R,
+extern "C" void shift_subtract_stack(unsigned int * L,
+unsigned int * R,
 float * out,
 int rows, int cols);
